pyproject.toml
[tool.poetry]
name = "training-projects"
version = "0.1.0"
description = ""
authors = ["pavlodubovyi <133531274+pavlodubovyi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
